rename, add test, add comments
