+++ examples/offload_compute/test_demo.py
@@ -1,0 +1,5 @@
+from .main import demo
+
+
+def test_demo():
+    demo()
